+++ Packages/com.yontalane.gridnav/Runtime/GridNavigator.cs
@@ -1,14 +1,15 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 
 namespace Yontalane.GridNav
 {
-    public class GridNavigator<T> where T : class, IGridNode
+    public class GridNavigator
     {
         #region Structs & Enums
         private struct VisitedNode
         {
-            public T node;
+            public Vector2Int coordinate;
             public int visited;
         }
 
@@ -31,9 +32,16 @@ namespace Yontalane.GridNav
         private Vector2Int m_startCoord;
         private Vector2Int m_endCoord;
         private readonly List<VisitedNode> m_path;
+        private Vector2Int m_gridSize;
+        private Func<Vector2Int, bool> m_nodeIsValid;
         #endregion
 
-        public GridNavigator() => m_path = new List<VisitedNode>();
+        public GridNavigator(Vector2Int gridSize, Func<Vector2Int, bool> nodeIsValid)
+        {
+            m_gridSize = gridSize;
+            m_nodeIsValid = nodeIsValid;
+            m_path = new List<VisitedNode>();
+        }
 
         /// <summary>
         /// The number of nodes along the path to reach the goal.
@@ -43,31 +51,31 @@ namespace Yontalane.GridNav
         /// <summary>
         /// The node at the given index along the path to reach the goal.
         /// </summary>
-        public T GetPathNode(int index) => index >= 0 && index < m_path.Count ? m_path[index].node : null;
+        public Vector2Int GetPathNode(int index) => m_path[index].coordinate;
 
         /// <summary>
         /// Find a path to the goal. Invokes delegate when finding is complete.
         /// </summary>
-        public void FindPath(Vector2Int start, Vector2Int end, T[,] gridNodes)
+        public void FindPath(Vector2Int start, Vector2Int end)
         {
             m_startCoord = start;
             m_endCoord = end;
-            GenerateGrid(gridNodes);
+            GenerateGrid();
             SetDistance();
             SetPath();
         }
 
-        private void GenerateGrid(T[,] gridNodes)
+        private void GenerateGrid()
         {
-            m_visitedNodes = new VisitedNode[gridNodes.GetLength(0), gridNodes.GetLength(1)];
+            m_visitedNodes = new VisitedNode[m_gridSize.x, m_gridSize.y];
 
-            for (int x = 0; x < gridNodes.GetLength(0); x++)
+            for (int x = 0; x < m_visitedNodes.GetLength(0); x++)
             {
-                for (int y = 0; y < gridNodes.GetLength(1); y++)
+                for (int y = 0; y < m_visitedNodes.GetLength(1); y++)
                 {
                     m_visitedNodes[x, y] = new VisitedNode()
                     {
-                        node = gridNodes[x, y],
+                        coordinate = new Vector2Int(x, y),
                         visited = -1
                     };
                 }
@@ -81,11 +89,11 @@ namespace Yontalane.GridNav
             int[] testArray = new int[m_visitedNodes.GetLength(0) * m_visitedNodes.GetLength(1)];
             for (int step = 1; step < testArray.Length; step++)
             {
-                foreach (VisitedNode obj in m_visitedNodes)
+                foreach (VisitedNode visitedNode in m_visitedNodes)
                 {
-                    if (obj.node != null && obj.visited == step - 1)
+                    if (m_nodeIsValid(visitedNode.coordinate) && visitedNode.visited == step - 1)
                     {
-                        TestFourDirections(obj.node.GetCoordinate(), step);
+                        TestFourDirections(visitedNode.coordinate, step);
                     }
                 }
             }
@@ -97,7 +105,7 @@ namespace Yontalane.GridNav
             Vector2Int coord = m_endCoord;
             List<VisitedNode> tempList = new List<VisitedNode>();
             m_path.Clear();
-            if (coord.x >= 0 && coord.x < m_visitedNodes.GetLength(0) && coord.y >= 0 && coord.y < m_visitedNodes.GetLength(1) && m_visitedNodes[m_endCoord.x, m_endCoord.y].node != null && m_visitedNodes[m_endCoord.x, m_endCoord.y].visited > 0)
+            if (coord.x >= 0 && coord.x < m_visitedNodes.GetLength(0) && coord.y >= 0 && coord.y < m_visitedNodes.GetLength(1) && m_nodeIsValid(m_visitedNodes[m_endCoord.x, m_endCoord.y].coordinate) && m_visitedNodes[m_endCoord.x, m_endCoord.y].visited > 0)
             {
                 m_path.Insert(0, m_visitedNodes[coord.x, coord.y]);
                 step = m_visitedNodes[coord.x, coord.y].visited - 1;
@@ -128,7 +136,7 @@ namespace Yontalane.GridNav
 
                 VisitedNode tempObj = FindClosest(m_visitedNodes[m_endCoord.x, m_endCoord.y], tempList);
                 m_path.Insert(0, tempObj);
-                coord = tempObj.node.GetCoordinate();
+                coord = tempObj.coordinate;
                 tempList.Clear();
             }
             OnComplete?.Invoke(true);
@@ -140,7 +148,7 @@ namespace Yontalane.GridNav
             {
                 for (int y = 0; y < m_visitedNodes.GetLength(1); y++)
                 {
-                    if (m_visitedNodes[x, y].node != null)
+                    if (m_nodeIsValid(m_visitedNodes[x, y].coordinate))
                     {
                         m_visitedNodes[x, y].visited = -1;
                     }
@@ -154,7 +162,7 @@ namespace Yontalane.GridNav
             switch (direction)
             {
                 case Direction.Up:
-                    if (coord.y + 1 < m_visitedNodes.GetLength(1) && m_visitedNodes[coord.x, coord.y + 1].node != null && m_visitedNodes[coord.x, coord.y + 1].visited == step)
+                    if (coord.y + 1 < m_visitedNodes.GetLength(1) && m_nodeIsValid(m_visitedNodes[coord.x, coord.y + 1].coordinate) && m_visitedNodes[coord.x, coord.y + 1].visited == step)
                     {
                         return true;
                     }
@@ -163,7 +171,7 @@ namespace Yontalane.GridNav
                         return false;
                     }
                 case Direction.Right:
-                    if (coord.x + 1 < m_visitedNodes.GetLength(0) && m_visitedNodes[coord.x + 1, coord.y].node != null && m_visitedNodes[coord.x + 1, coord.y].visited == step)
+                    if (coord.x + 1 < m_visitedNodes.GetLength(0) && m_nodeIsValid(m_visitedNodes[coord.x + 1, coord.y].coordinate) && m_visitedNodes[coord.x + 1, coord.y].visited == step)
                     {
                         return true;
                     }
@@ -172,7 +180,7 @@ namespace Yontalane.GridNav
                         return false;
                     }
                 case Direction.Down:
-                    if (coord.y - 1 >= 0 && m_visitedNodes[coord.x, coord.y - 1].node != null && m_visitedNodes[coord.x, coord.y - 1].visited == step)
+                    if (coord.y - 1 >= 0 && m_nodeIsValid(m_visitedNodes[coord.x, coord.y - 1].coordinate) && m_visitedNodes[coord.x, coord.y - 1].visited == step)
                     {
                         return true;
                     }
@@ -181,7 +189,7 @@ namespace Yontalane.GridNav
                         return false;
                     }
                 case Direction.Left:
-                    if (coord.x - 1 >= 0 && m_visitedNodes[coord.x - 1, coord.y].node != null && m_visitedNodes[coord.x - 1, coord.y].visited == step)
+                    if (coord.x - 1 >= 0 && m_nodeIsValid(m_visitedNodes[coord.x - 1, coord.y].coordinate) && m_visitedNodes[coord.x - 1, coord.y].visited == step)
                     {
                         return true;
                     }
@@ -219,7 +227,7 @@ namespace Yontalane.GridNav
             {
                 if (coord.y >= 0 && coord.y < m_visitedNodes.GetLength(1))
                 {
-                    if (m_visitedNodes[coord.x, coord.y].node != null)
+                    if (m_nodeIsValid(m_visitedNodes[coord.x, coord.y].coordinate))
                     {
                         m_visitedNodes[coord.x, coord.y].visited = step;
                     }
@@ -233,9 +241,9 @@ namespace Yontalane.GridNav
             int closestIndex = 0;
             for (int i = 0; i < list.Count; i++)
             {
-                if (list[i].node != null)
+                if (m_nodeIsValid(list[i].coordinate))
                 {
-                    float distance = Vector2.Distance(targetLocation.node.GetCoordinate(), list[i].node.GetCoordinate());
+                    float distance = Vector2.Distance(targetLocation.coordinate, list[i].coordinate);
                     if (distance < closestDistance)
                     {
                         closestDistance = distance;