+++ Packages/com.yontalane.gridnav/Runtime/IGridNode.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-namespace Yontalane.GridNav
-{
-    public interface IGridNode
-    {
-        public Vector2Int GetCoordinate();
-    }
-}